Fixes a few SGI compiler warnings..
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@2436 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/gtk1/dc.h
@@ -88,7 +88,7 @@ public:
   {
       CrossHair(pt.x, pt.y);
   }
-  virtual void DrawArc( long x1, long y1, long x2, long y2, double xc, double yc );
+  virtual void DrawArc( long x1, long y1, long x2, long y2, long xc, long yc );
   inline void DrawArc(const wxPoint& pt1, const wxPoint& pt2, const wxPoint& centre)
   {
       DrawArc(pt1.x, pt1.y, pt2.x, pt2.y, centre.x, centre.y);
