+++ include/wx/gtk1/dcclient.h
@@ -45,7 +45,7 @@ public:
 
   virtual void DrawLine( long x1, long y1, long x2, long y2 );
   virtual void CrossHair( long x, long y );
-  virtual void DrawArc( long x1, long y1, long x2, long y2, double xc, double yc );
+  virtual void DrawArc( long x1, long y1, long x2, long y2, long xc, long yc );
   virtual void DrawEllipticArc( long x, long y, long width, long height, double sa, double ea );
   virtual void DrawPoint( long x, long y );
